Portable Text Images loading correctly
Correctd the graphql query in gatsby-node.js to correctly query the
portable text raw data.

Added a serializer component that is able to correctly handle images
inserted into portable text.

Added a Figure component that queries graphql for the image data and
loads in with gatsby dynamic images.

fixed the box shadow on the products page to transition and removed
border.

removed the serializer from post.js and added it to it's own file.

Fixed some HTML tags in laylout.js and products.js for better semantics

added card shadow variables to layout.css