Created dropdowns for sub-pages
added a menuItem component that handles programmatically generated
links for pages and the logic for creating dropdowns with links to sub
pages.

Created a nav.css file to keep all of the nav related css in one place

add the nav.css file to layout and replaced the programmatically
generated links with the MenuItem.js component.

removed the header nav related css from the layout.css file